+++ runtests.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python3
+#!/usr/bin/python3 -u
 """
 Run the test suite on multiple Python versions.